3 - Dynamic Website/backend/requirements.txt
fastapi
uvicorn
strands-agents[litellm]>=0.1.0
strands-agents-tools>=0.1.0
strands-agents-builder>=0.1.0
pydantic
python-dotenv
strands-agents[openai]
